+++ lib/layouts/Article/Article.tsx
@@ -1,5 +1,4 @@
 import React from 'react'
-import { proseStyles } from '../../utils/globalClasses'
 
 export interface ArticleProps {
   children?: React.ReactNode
@@ -9,10 +8,8 @@ export interface ArticleProps {
 export const Article = ({ children, content }: ArticleProps) => {
   return (
     <>
-      {content && (
-        <article className={`cu-article ${proseStyles.base}`} dangerouslySetInnerHTML={{ __html: content }} />
-      )}
-      {!content && <article className={`cu-article ${proseStyles.base}`}>{children}</article>}
+      {content && <article className="cu-article" dangerouslySetInnerHTML={{ __html: content }} />}
+      {!content && <article className="cu-article">{children}</article>}
     </>
   )
 }
