+++ lib/main.ts
@@ -66,7 +66,7 @@ export { Body } from './layouts/Body/Body'
 export { Column } from './layouts/Column/Column'
 export { Container } from './layouts/Container/Container'
 export { Main } from './layouts/Main/Main'
-export { Section } from './layouts/Section/Section'
+export { Section } from './_deprecated/Section/Section.js'
 export { StackedList } from './layouts/StackedList/StackedList'
 
 // Javascript Files
